refactoring uptime-server
