+++ practice/uptime-server/src/types.rs
@@ -1,0 +1,3 @@
+pub type ServiceAlias = String;
+pub type Address = String;
+
